[Flavia/Jo] Issue #69: Refatoração do componente Header
