+++ src/pages/_layout/Layout.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import grey from '@material-ui/core/colors/grey';
 import makeStyles from '@material-ui/core/styles/makeStyles';
-import Header from './Header';
+import Header from '../../components/Header/Header';
 import Footer from '../../components/Footer/Footer';
+import Menu from '../../components/Menu/Menu';
 
 const useStyle = makeStyles(() => ({
   layout: {
@@ -22,6 +23,7 @@ const Layout = (props) => {
   return (
     <div className={classes.layout}>
       <Header />
+      <Menu />
       {children}
       <Footer />
     </div>
